requirements.txt
streamlit>=1.45.0
google-cloud-bigquery>=3.35.0
google-genai>=0.9.0
pandas>=2.2.0
sqlparse>=0.5.0
plotly>=5.24.0
networkx>=3.3
python-dotenv>=1.0.0
db-dtypes>=1.2.0
pyarrow>=15.0.0